Change test to use and rename ga_pygad
